+++ R/create_mixsiar_envir.R
@@ -8,4 +8,4 @@
 #'
 mixsiar <- new.env()
 # Added only to pass R CMD check with 0 NOTEs
-if(getRversion() >= "2.15.1")  utils::globalVariables(c("p.global","ilr.global","ilr.fac1","ilr.fac2","fac1.sig","fac2.sig","ind.sig","..scaled..","p.fac1","p.fac2","p.ind","sources","x","position_jitter","scolour","xmin","xmax","label","y","ymin","ymax"))
+if(getRversion() >= "2.15.1")  utils::globalVariables(c("p.global","ilr.global","ilr.fac1","ilr.fac2","fac1.sig","fac2.sig","ind.sig","..scaled..","p.fac1","p.fac2","p.ind","sources","x","position_jitter","scolour","xmin","xmax","label","y","ymin","ymax","low","high"))